requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.6.1
slowapi==0.1.8
tiktoken~=0.9.0
protobuf~=4.25.6
tenacity~=9.1.2